+++ CHANGELOG.md
@@ -9,11 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - option to update `minimumStakingTokenPercentage`.
 - `initialBlockHeights` to set the initial block-height per shard
 - fetch VerifierRegistry contract from ContractRegistry
-- Proposals are made for specyfic `block-height`
+- Proposals are made for specific `block-height`
+- Reveals are made for specific `block-height`
 
 ### Changed
-- smart contracts are complient with solidity 0.5
-- unit tests are complient with web3 1.0
+- smart contracts are compliant with solidity 0.5
+- unit tests are compliant with web3 1.0
 - when verifier proposing, `Chain` checks, if verifier is active instead
   of if it is only created.
 - split Chain contract into two separate contracts (storage and manager)